backend/requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
PyMuPDF>=1.23.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
